lint fix + others
